Feed tweets

+++ src/main/java/br/com/posterius/spring_security/dto/FeedDto.java
@@ -2,6 +2,6 @@ package br.com.posterius.spring_security.dto;
 
 import java.util.List;
 
-public record FeedDto(List<FeedItemDto> feedItens, int pages, int pageSize, int totalPages, int totalElements) {
+public record FeedDto(List<FeedItemDto> feedItens, int pages, int pageSize, int totalPages, long totalElements) {
 
 }
